fix: address PR review — move to flat file, tighten types, add cross-ref
- Move src/lib/utils/unwrap-content.ts → src/lib/unwrap-content.ts to
  match house convention (flat files in src/lib/)
- Tighten type cast: use typeof checks instead of bare 'as string' cast
- Use paragraph break joiner (\n\n) for multi-block content
- Add cross-reference comment to electron/agent/app-runtime.ts
  normalizeAssistantText() so future contributors know both exist
